+++ RELEASES.md
@@ -1,3 +1,5 @@
+* Add `Float::get_significand` method.
+
 Version 1.0.1 (2018-03-10)
 ==========================